chore: 모듈 구조 수정
