+++ src/app.module.ts
@@ -2,8 +2,6 @@ import { Module, ValidationPipe } from '@nestjs/common';
 import { ConfigModule } from '@nestjs/config';
 import { APP_PIPE } from '@nestjs/core';
 import { AuthModule } from './auth/auth.module';
-import { ProductModule } from './product/product.module';
-import { SellerModule } from './seller/seller.module';
 
 @Module({
   imports: [
@@ -12,8 +10,6 @@ import { SellerModule } from './seller/seller.module';
       cache: true,
       isGlobal: true,
     }),
-    ProductModule,
-    SellerModule,
   ],
   providers: [
     {
