--- src/product/product.module.ts
@@ -1,9 +1,0 @@
-import { Module } from '@nestjs/common';
-import { ProductController } from './product.controller';
-import { ProductService } from './product.service';
-
-@Module({
-  controllers: [ProductController],
-  providers: [ProductService]
-})
-export class ProductModule {}
